+++ HISTORY.md
@@ -3,6 +3,7 @@
 `NEW` FormField can set its own mode different from Form now.
 `NEW` add GridFormField.
 `FIXED` when props jsxvalues change, form cannot handle correctly.
+`FIXED` when props jsxmode change, form cannot handle correctly.
 
 
 ## 1.0.6